+++ volmdlr/shells.py
@@ -157,9 +157,61 @@ class OpenShell3D(volmdlr.core.CompositePrimitive3D):
             self._vertices_graph = vertices_graph
         return self._vertices_graph
 
-    def faces_graph(self):
+    def _faces_graph_search_bridges(self, graph, components, face_vertices):
         """
-        Gets the shells faces graph using networkx.
+        Search for neighboring faces in the connected components to fix the graph.
+
+        To make the shell faces' topology graph, we search faces that share the same vertices.
+        Sometimes, in very specific cases, it can occur that two faces are neighbors of each other
+        but their vertices aren't coincident.
+
+        This method tries to find those cases, by searching bridges, which are edges that connect two components.
+        The search is performed by checking if any vertices of one face in one component lie on the contours of
+        another face in the other component. If such a connection is found, the two faces are considered neighbors.
+        Once a connection is found between a pair of components we break the execution and proceed to
+        the next pair of components. This behavior, in the average, avoids the need to verify all possible combinations
+
+        :param graph: The graph representing the faces' topology.
+        :type graph: nx.Graph
+        :param components: A list of sets, where each set contains the indices of faces in a connected component.
+        :type components: list
+        :param face_vertices: A dictionary mapping face indices to their corresponding vertex indices.
+        :type face_vertices: dict
+        :return: The updated graph with no disconnected components.
+        """
+        stack = components.copy()
+        found = False
+
+        def check_faces(face, other_face_id):
+            for point_id in face_vertices[other_face_id]:
+                point = self.vertices_points[point_id]
+                if face.outer_contour3d.point_over_contour(point):
+                    return True
+                if face.inner_contours3d:
+                    for inner_contour in face.inner_contours3d:
+                        if inner_contour.point_over_contour(point):
+                            return True
+            return False
+
+        while stack:
+            group = stack.pop(0)
+            for face_id_i in group:
+                face_i = self.faces[face_id_i]
+                for other_group in stack:
+                    for face_id_j in other_group:
+                        if check_faces(face_i, face_id_j):
+                            found = True
+                            graph.add_edge(face_id_i, face_id_j)
+                            for face_id_k in graph.neighbors(face_id_j):
+                                if check_faces(face_i, face_id_k):
+                                    graph.add_edge(face_id_i, face_id_k)
+                        if found:
+                            break
+        return graph
+
+    def faces_graph(self, verify_connected_components=True):
+        """
+        Gets the shells faces topology graph using networkx.
 
         :return: return a networkx graph for a shell faces.
         """
@@ -167,30 +219,25 @@ class OpenShell3D(volmdlr.core.CompositePrimitive3D):
             graph = nx.Graph()
             vertice_faces = {}
             face_vertices = {}
-            vertice_edges = {}
-            face_edges = {}
             for face_index, face in enumerate(self.faces):
-                outer_contour = face.outer_contour3d
-                for edge in outer_contour.primitives:
+                face_contour_primitives = face.outer_contour3d.primitives
+                for inner_contour in face.inner_contours3d:
+                    face_contour_primitives.extend(inner_contour.primitives)
+                for edge in face_contour_primitives:
                     start_index = volmdlr.core.get_point_index_in_list(edge.start, self.vertices_points)
                     vertice_faces.setdefault(start_index, set()).add(face_index)
-                    vertice_edges.setdefault(start_index, set()).add(edge)
-                    face_edges.setdefault(face_index, set()).add(edge)
                     face_vertices.setdefault(face_index, set()).add(start_index)
-            for i, face_i in enumerate(self.faces):
-                face_i_edges = face_edges[i]
+            for i, _ in enumerate(self.faces):
                 face_i_vertices = face_vertices[i]
                 for vertice in face_i_vertices:
                     connected_faces = vertice_faces[vertice]
                     connected_faces.discard(i)
                     for j in connected_faces:
-                        face_j_edges = face_edges[j]
-                        edges_j_set = face_j_edges.intersection(vertice_edges[vertice])
-                        for edge_i in face_i_edges:
-                            for edge_j in edges_j_set:
-                                if edge_i.get_shared_section(edge_j):
-                                    graph.add_edge(i, j)
-
+                        graph.add_edge(i, j)
+            if verify_connected_components:
+                components = list(nx.connected_components(graph))
+                if len(components) > 1:
+                    graph = self._faces_graph_search_bridges(graph, components, face_vertices)
             self._faces_graph = graph
         return self._faces_graph
 
@@ -804,7 +851,7 @@ class OpenShell3D(volmdlr.core.CompositePrimitive3D):
         Defines a List of separated OpenShell3D from a list of faces, based on the faces graph.
         """
         class_ = getattr(sys.modules[__name__], cls._from_face_class)
-        graph = class_(faces).faces_graph()
+        graph = class_(faces).faces_graph(verify_connected_components=False)
         components = [graph.subgraph(c).copy() for c in nx.connected_components(graph)]
 
         shells_list = []
@@ -812,7 +859,6 @@ class OpenShell3D(volmdlr.core.CompositePrimitive3D):
             faces_list = []
             for n_index in graph_i.nodes:
                 faces_list.append(faces[n_index])
-
             shells_list.append(class_(faces_list))
 
         return shells_list