+++ CHANGELOG.md
@@ -57,6 +57,7 @@ All adequations have been done for all tests and existing scripts.
 - bspline_compiled: refactor binomial_coefficient for performance.
 - Improve step translator.
 - Delete inplace methods: rotation, translation and frame_mapping
+- OpenShell3D: faces_graph.
 
 
 ### Changed